scripts: Only update software on list change
Previously if anything in packages.yaml or ~/.config/bmatheny/packages.yaml
changed, it would trigger a run of all software install scripts.

With this change, we generate hashes specifically for the relevant portion of
the config. So a run of the software installer is only triggered if that
software installers list changed.