scripts: shared code, user specified brews
This commit adds two things; shared code for scripts, and the ability for users
to provide their own home brew file.

For shared code, it turned out it's challenging to allow for code reuse from
within the '.chezmoiscripts' directory for a variety of reasons. What ended up
making the most sense given the issues I ran into was; code is captured in a
support directory, and scripts wanting to use that code follows the
"initialize_support" strategy found in the brews scripts. Essentially most of
the magic here is due to me wanting to allow for chezmoi templates to be run
both as a standalone script for testing, and also as a normal chezmoi script.

For users providing their own home brew file, just put a
~/.config/bmatheny/packages.yaml file in place with your packages you would like
installed. I'll almost immediately move this to Brewfile format for a few
reasons, but that's ok since nobody uses my repo anyways.